[UPDATE] fix some announcement not deleteing

+++ src/components/button/announcementButtons/announcementSubmit.js
@@ -59,7 +59,9 @@ module.exports = {
       )
     );
 
-    await existingThread.delete();
+    if (existingThread) {
+      await existingThread.delete();
+    }
 
     await interaction.message.delete();
